# cache: remove prefetch closure